optimize IAM online API

+++ super-devops-iam/super-devops-iam-common/src/main/java/com/wl4g/devops/iam/common/web/GenericApiController.java
@@ -26,8 +26,6 @@ import com.wl4g.devops.iam.common.web.model.SessionAttributeModel;
 import com.wl4g.devops.iam.common.web.model.SessionDestroyModel;
 import com.wl4g.devops.iam.common.web.model.SessionQueryModel;
 import com.wl4g.devops.support.cache.ScanCursor;
-import com.wl4g.devops.support.cache.ScanCursor.CursorWrapper;
-import org.apache.shiro.subject.PrincipalCollection;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -36,16 +34,19 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.io.Serializable;
+import java.util.Collection;
 import java.util.List;
 
+import static com.wl4g.devops.support.cache.ScanCursor.CursorWrapper.*;
 import static com.wl4g.devops.common.constants.IAMDevOpsConstants.BEAN_DELEGATE_MSG_SOURCE;
 import static com.wl4g.devops.common.constants.IAMDevOpsConstants.URI_S_API_V1_SESSION;
 import static com.wl4g.devops.iam.common.web.model.SessionAttributeModel.CursorIndex;
 import static com.wl4g.devops.iam.common.web.model.SessionAttributeModel.SessionAttribute;
 import static java.util.Objects.nonNull;
 import static java.util.stream.Collectors.toList;
+import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.apache.shiro.web.subject.support.DefaultWebSubjectContext.AUTHENTICATED_SESSION_KEY;
-import static org.apache.shiro.web.subject.support.DefaultWebSubjectContext.PRINCIPALS_SESSION_KEY;
+import static org.springframework.util.CollectionUtils.isEmpty;
 
 /**
  * Generic abstract API controller.
@@ -77,8 +78,8 @@ public abstract class GenericApiController extends BaseController {
 	protected IamSessionDAO sessionDAO;
 
 	/**
-	 * Iterative scan gets the list of access sessions (including all clients and
-	 * authenticated and uncertified sessions).</br>
+	 * Iterative scan gets the list of access sessions (including all clients
+	 * and authenticated and uncertified sessions).</br>
 	 * <p>
 	 * For example response:
 	 *
@@ -137,15 +138,20 @@ public abstract class GenericApiController extends BaseController {
 			log.info("Get sessions by <= {}", query);
 		}
 
-		// Parse cursor.
-		CursorWrapper cursor = CursorWrapper.parse(query.getCursor());
-		// Do scan access sessions all.
-		ScanCursor<IamSession> sc = sessionDAO.getAccessSessions(cursor, query.getLimit());
-		List<SessionAttribute> sas = sc.readValues().stream().map(s -> wrapSessionAttribute(s)).collect(toList());
-
-		// Setup response attributes.
-		CursorIndex index = new CursorIndex(sc.getCursor().getCursorString(), sc.getCursor().getHasNext());
-		resp.setData(new SessionAttributeModel(index, sas));
+		// Priority search principal.
+		if (!isBlank(query.getPrincipal())) {
+			Collection<IamSession> ss = sessionDAO.getAccessSessions(query.getPrincipal());
+			List<SessionAttribute> sas = ss.stream().map(s -> wrapSessionAttribute(s)).collect(toList());
+			resp.setData(new SessionAttributeModel(new CursorIndex(false), sas));
+		} else {
+			// Do scan sessions all.
+			ScanCursor<IamSession> sc = sessionDAO.getAccessSessions(parse(query.getCursor()), query.getLimit());
+			// Convert to SessionAttribute.
+			List<SessionAttribute> sas = sc.readValues().stream().map(s -> wrapSessionAttribute(s)).collect(toList());
+			// Setup response attributes.
+			CursorIndex index = new CursorIndex(sc.getCursor().getCursorString(), sc.getCursor().getHasNext());
+			resp.setData(new SessionAttributeModel(index, sas));
+		}
 
 		if (log.isInfoEnabled()) {
 			log.info("Get sessions => {}", resp.asJson());
@@ -161,15 +167,22 @@ public abstract class GenericApiController extends BaseController {
 	 * @throws Exception
 	 */
 	@PostMapping(path = URI_S_API_V1_SESSION)
-	public RespBase<?> destroySession(@Validated @RequestBody SessionDestroyModel destroy) throws Exception {
+	public RespBase<?> destroySessions(@Validated @RequestBody SessionDestroyModel destroy) throws Exception {
 		RespBase<String> resp = RespBase.create();
 		if (log.isInfoEnabled()) {
 			log.info("Destroy sessions by <= {}", destroy);
 		}
 
-		// Destroy sessions.
-		for (Serializable sessionId : destroy.getSessionIds()) {
-			sessionDAO.delete(new IamSession(sessionId));
+		// Destroy with sessionIds.
+		if (!isEmpty(destroy.getSessionIds())) {
+			for (Serializable sessionId : destroy.getSessionIds()) {
+				sessionDAO.delete(new IamSession(sessionId));
+			}
+		}
+
+		// Destroy with principal.
+		if (!isBlank(destroy.getPrincipal())) {
+			sessionDAO.removeAccessSession(destroy.getPrincipal());
 		}
 
 		if (log.isInfoEnabled()) {
@@ -259,10 +272,7 @@ public abstract class GenericApiController extends BaseController {
 		}
 
 		// Authentication principal.
-		PrincipalCollection principals = (PrincipalCollection) session.getAttribute(PRINCIPALS_SESSION_KEY);
-		if (nonNull(principals) && !principals.isEmpty()) {
-			sa.setPrincipal(principals.getPrimaryPrincipal());
-		}
+		sa.setPrincipal(session.getPrimaryPrincipal());
 
 		return sa;
 	}
